Split execution into separate function and file
